Changed from brush to opacity
