+++ src/settings-ui/Microsoft.PowerToys.Settings.UI/App.xaml
@@ -17,13 +17,7 @@
                 <ResourceDictionary Source="/Styles/Page.xaml"/>
                 <ResourceDictionary Source="/Styles/Button.xaml"/>
             </ResourceDictionary.MergedDictionaries>
-            <converters:ModuleEnabledToForegroundConverter x:Key="ModuleEnabledToForegroundConverter"/>
-
-            <SolidColorBrush x:Key="DarkForegroundDisabledBrush">#66FFFFFF</SolidColorBrush>
-            <SolidColorBrush x:Key="DarkForegroundBrush">#FFFFFFFF</SolidColorBrush>
-
-            <SolidColorBrush x:Key="LightForegroundDisabledBrush">#66000000</SolidColorBrush>
-            <SolidColorBrush x:Key="LightForegroundBrush">#FF000000</SolidColorBrush>
+            <converters:ModuleEnabledToOpacityConverter x:Key="ModuleEnabledToOpacityConverter"/>
         </ResourceDictionary>
     </Application.Resources>
 </xaml:XamlApplication>
